scripts/requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0

# Machine Learning
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0

# LLM
groq>=0.4.0

# Type checking (optional)
# mypy>=1.0.0
